optimised code

+++ app/controllers/Teachers/DeleteBook.php
@@ -3,20 +3,14 @@
 
 class DeleteBook{
     public function get(){
-        if(isset($_SESSION['uid'])){
-            if($_SESSION['type']=='teacher'){
+        if(isset($_SESSION['uid']) && $_SESSION['type']=='teacher'){
                 \Model\Books::deleteById($_GET['id']);
-                $host  = $_SERVER['HTTP_HOST'];
-                header("Location: http://$host/teacher/home");
-            }
-            else{
-                $host  = $_SERVER['HTTP_HOST'];
-                header("Location: http://$host/");
-            }
+                
+                header("Location: /teacher/home");
         }
         else{
-            $host  = $_SERVER['HTTP_HOST'];
-            header("Location: http://$host/");
+            
+            header("Location: /");
         }
     }
 }
